Jeremy  - 更新簡中語言包

See merge request Lin/lss_vue_enigma!37

+++ src/views/category-management/Main.vue
@@ -65,6 +65,7 @@ import { list_product_category, create_product_category, update_product_category
 import { onMounted, ref, computed } from "vue";
 import { useLSSSellerLayoutStore } from "@/stores/lss-seller-layout"
 import dom from "@left4code/tw-starter/dist/js/dom";
+import i18n from "@/locales/i18n"
 
 const layoutStore = useLSSSellerLayoutStore()
 const listItems = ref([])
@@ -129,24 +130,21 @@ const showEditModal = item=>{
 
 
 function deleteCategory(name) {
-    delete_product_category(name).then(
-        response => {
-            list();
-        }
-    )
+    let yes = confirm(`${i18n.global.t('stock.category_manage.confirm_delete')}`)
+    if (yes) delete_product_category(name).then(res => { list() } )
     hideDropDown()
 }
 
 function closeAlert() {
-		if (saved.value === true) {
-			showModal.value = false;
-			layoutStore.notification.showMessageToast("Save Success")
-		} else {
-			showModal.value = false;
-			layoutStore.alert.showMessageToast("Change Not Saved")
-		}
-		saved.value = false
-        oldCategory.value = ''
-        categoryName.value = ''
-	}
+    if (saved.value === true) {
+        showModal.value = false;
+        layoutStore.notification.showMessageToast(i18n.global.t('stock.category_manage.save_success'))
+    } else {
+        showModal.value = false;
+        layoutStore.alert.showMessageToast(i18n.global.t('stock.category_manage.not_saved'))
+    }
+    saved.value = false
+    oldCategory.value = ''
+    categoryName.value = ''
+}
 </script>
